fix: FOSSA dependency issue
