+++ operate/webapp/pom.xml
@@ -298,10 +298,6 @@
       <artifactId>spring-test</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>io.camunda</groupId>
-      <artifactId>configuration</artifactId>
-    </dependency>
   </dependencies>
 
   <build>
