+++ spring-utils/src/main/java/io/camunda/spring/utils/ConditionalOnWebappUiEnabled.java
@@ -5,16 +5,16 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.configuration.conditions;
+package io.camunda.spring.utils;
 
-import com.google.common.collect.Sets;
-import io.camunda.configuration.conditions.ConditionalOnWebappEnabled.OnWebappEnabledCondition;
-import io.camunda.configuration.conditions.ConditionalOnWebappUiEnabled.OnWebappUiEnabledCondition;
+import io.camunda.spring.utils.ConditionalOnWebappEnabled.OnWebappEnabledCondition;
+import io.camunda.spring.utils.ConditionalOnWebappUiEnabled.OnWebappUiEnabledCondition;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.util.HashSet;
 import java.util.Set;
 import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.Conditional;
@@ -36,9 +36,10 @@ public @interface ConditionalOnWebappUiEnabled {
 
     @Override
     protected Set<String> getRequiredProperties(final String webappName) {
-      return Sets.union(
-          super.getRequiredProperties(webappName),
-          Set.of("camunda.webapps." + webappName + ".ui-enabled"));
+      final Set<String> result = new HashSet<>();
+      result.addAll(super.getRequiredProperties(webappName));
+      result.add("camunda.webapps." + webappName + ".ui-enabled");
+      return result;
     }
   }
 }
