+++ spring-utils/src/test/java/io/camunda/spring/utils/ConditionalOnWebappUiEnabledTest.java
@@ -5,12 +5,12 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.configuration.conditions;
+package io.camunda.spring.utils;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import io.camunda.configuration.conditions.ConditionalOnWebappEnabled.OnWebappEnabledCondition;
+import io.camunda.spring.utils.ConditionalOnWebappUiEnabled.OnWebappUiEnabledCondition;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -18,20 +18,21 @@ import org.springframework.context.annotation.ConditionContext;
 import org.springframework.core.env.Environment;
 import org.springframework.core.type.AnnotatedTypeMetadata;
 
-public class ConditionalOnWebappEnabledTest {
+public class ConditionalOnWebappUiEnabledTest {
 
   private static final String TEST_WEBAPP_NAME = "testwebapp";
   private static final String LEGACY_PROPERTY = "camunda." + TEST_WEBAPP_NAME + ".webapp-enabled";
   private static final String UNIFIED_PROPERTY = "camunda.webapps." + TEST_WEBAPP_NAME + ".enabled";
+  private static final String UI_PROPERTY = "camunda.webapps." + TEST_WEBAPP_NAME + ".ui-enabled";
 
-  private OnWebappEnabledCondition condition;
+  private OnWebappUiEnabledCondition condition;
   private ConditionContext context;
   private Environment environment;
   private AnnotatedTypeMetadata metadata;
 
   @BeforeEach
   void setup() {
-    condition = new OnWebappEnabledCondition();
+    condition = new OnWebappUiEnabledCondition();
     context = mock(ConditionContext.class);
     environment = mock(Environment.class);
     metadata = mock(AnnotatedTypeMetadata.class);
@@ -42,10 +43,11 @@ public class ConditionalOnWebappEnabledTest {
   }
 
   @Test
-  void testShouldMatchwhenAllPropertiesAreTrue() {
+  void testShouldMatchWhenAllPropertiesAreTrue() {
     // given
     when(environment.getProperty(LEGACY_PROPERTY, Boolean.class, true)).thenReturn(true);
     when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(true);
+    when(environment.getProperty(UI_PROPERTY, Boolean.class, true)).thenReturn(true);
 
     // when
     final boolean result = condition.matches(context, metadata);
@@ -57,8 +59,9 @@ public class ConditionalOnWebappEnabledTest {
   @Test
   void testShouldNotMatchWhenLegacyPropertyIsFalse() {
     // given
-    when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(true);
     when(environment.getProperty(LEGACY_PROPERTY, Boolean.class, true)).thenReturn(false);
+    when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(true);
+    when(environment.getProperty(UI_PROPERTY, Boolean.class, true)).thenReturn(true);
 
     // when
     final boolean result = condition.matches(context, metadata);
@@ -70,8 +73,23 @@ public class ConditionalOnWebappEnabledTest {
   @Test
   void testShouldNotMatchWhenUnifiedPropertyIsFalse() {
     // given
-    when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(false);
     when(environment.getProperty(LEGACY_PROPERTY, Boolean.class, true)).thenReturn(true);
+    when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(false);
+    when(environment.getProperty(UI_PROPERTY, Boolean.class, true)).thenReturn(true);
+
+    // when
+    final boolean result = condition.matches(context, metadata);
+
+    // then
+    assertThat(result).isFalse();
+  }
+
+  @Test
+  void testShouldNotMatchWhenUIPropertyIsFalse() {
+    // given
+    when(environment.getProperty(LEGACY_PROPERTY, Boolean.class, true)).thenReturn(true);
+    when(environment.getProperty(UNIFIED_PROPERTY, Boolean.class, true)).thenReturn(true);
+    when(environment.getProperty(UI_PROPERTY, Boolean.class, true)).thenReturn(false);
 
     // when
     final boolean result = condition.matches(context, metadata);
