+++ tasklist/data-generator/pom.xml
@@ -16,17 +16,17 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>configuration</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>io.camunda</groupId>
       <artifactId>tasklist-common</artifactId>
     </dependency>
 
     <dependency>
       <groupId>io.camunda</groupId>
       <artifactId>webapps-schema</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>camunda-spring-utils</artifactId>
     </dependency>
 
     <!-- SPRING -->
