+++ tasklist/webapp/pom.xml
@@ -371,11 +371,6 @@
       <artifactId>camunda-security-protocol</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>io.camunda</groupId>
-      <artifactId>configuration</artifactId>
-    </dependency>
-
   </dependencies>
 
   <build>
