+++ tasklist/webapp/src/main/java/io/camunda/tasklist/WebappModuleConfiguration.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.tasklist;
 
-import io.camunda.configuration.conditions.ConditionalOnWebappEnabled;
+import io.camunda.spring.utils.ConditionalOnWebappEnabled;
 import jakarta.annotation.PostConstruct;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
